PBRT.jl

Documentation for PBRT.jl

PBRT.at — Method

at(r::ray, t::Float64)

Return the point at t along the ray r. Using the formula

PBRT.create_super_triangle — Method

Creats a super Triangle that bounds a Face

PBRT.foo — Method

foo()

Return a greeting message.

PBRT.hit! — Method

Slab method for AABB intersection

PBRT.in_circumcircle — Method

Function to check if a point is inside the circumcircle of a Triangle

PBRT.not_shared — Method

Function to check if an edge is not shared by an array of triangles

PBRT.reader — Method

Wavefront .obj file parser

PBRT.refract — Method

refract(uv::Vector{Float64}, n::Vector{Float64}, r::Float64)::Vector{Float64}

Refract a vector uv through a normal n with a refractive index r. Uses Snell's Law to calculate the refracted vector.

PBRT.triangulation — Method

Boywer-Watson triangulation algorithm